requirements.txt
annotated-types==0.6.0
anyio==3.7.1
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
distro==1.8.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
gspread==5.12.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
oauthlib==3.2.2
openai==1.3.4
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.1
pydantic_core==2.14.3
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
sniffio==1.3.0
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.1.0
